feat: Add sm_enum_msgs plus refactoring, bump to v0.5.0
The file sm_enum_msgs.rs is an example where a boxed enum based message
protocols are defined and are passed "as-is" to the state machines. This
is how my current HSM's are implemented and I wanted to be able to have
that "old" style in this project for comparison.

In doing so I've renamed the sm_xxx_messages.rs files to sm_xxx_msgs_any.rs
so it's easy to distinguish between the "as-is" style and "Any" style.
Plus renaming trait ProcessMsg to ProcessMsgAny and its method to
process_msg_any. And renaming `Message = dyn Any` to `MsgAny = dyn Any`.